Remove block index from network processing
