+++ src/validation.cpp
@@ -42,6 +42,7 @@
 #include <validationinterface.h>
 #include <warnings.h>
 #include <libethcore/ABI.h>
+#include <net_processing.h>
 
 #include <serialize.h>
 #include <pubkey.h>
@@ -6788,69 +6789,9 @@ std::string exceptedMessage(const dev::eth::TransactionException& excepted, cons
     return message;
 }
 
-bool NeedToEraseBlockIndex(const CBlockIndex *pindex, const CBlockIndex *pindexCheck)
-{
-    if(!chainActive.Contains(pindex))
-    {
-        if(pindex->nHeight <= pindexCheck->nHeight) return true;
-        const CBlockIndex *pindexBlock = pindex;
-        while(pindexBlock)
-        {
-           pindexBlock = pindexBlock->pprev;
-           if(pindexBlock->nHeight == pindexCheck->nHeight) return pindexBlock != pindexCheck;
-        }
-    }
-    return false;
-}
-
-bool RemoveBlockIndex(CBlockIndex *pindex)
+bool RemoveStateBlockIndex(CBlockIndex *pindex)
 {
     return g_chainstate.RemoveBlockIndex(pindex);
-}
-
-void CleanBlockIndex()
-{
-    unsigned int cleanTimeout = gArgs.GetArg("-cleanblockindextimeout", DEFAULT_CLEANBLOCKINDEXTIMEOUT) * 1000;
-    if(cleanTimeout == 0) cleanTimeout = DEFAULT_CLEANBLOCKINDEXTIMEOUT * 1000;
-
-    while(!ShutdownRequested())
-    {
-        {
-            LOCK(cs_main);
-
-            std::vector<uint256> indexNeedErase;
-
-            const CBlockIndex *pindexCheck = chainActive[chainActive.Height() - nCheckpointSpan -1];
-
-            if(!IsInitialBlockDownload() && pindexCheck)
-            {
-                for (BlockMap::iterator it=mapBlockIndex.begin(); it!=mapBlockIndex.end(); it++)
-                {
-                    CBlockIndex *pindex = (*it).second;
-                    if(NeedToEraseBlockIndex(pindex, pindexCheck))
-                    {
-                        indexNeedErase.push_back(pindex->GetBlockHash());
-                    }
-                }
-
-                for(uint256 blockHash : indexNeedErase)
-                {
-                    BlockMap::iterator it=mapBlockIndex.find(blockHash);
-                    if(it!=mapBlockIndex.end())
-                    {
-                        CBlockIndex *pindex = (*it).second;
-                        if(RemoveBlockIndex(pindex))
-                        {
-                            delete pindex;
-                            mapBlockIndex.erase(it);
-                        }
-                    }
-                }
-            }
-        }
-
-        MilliSleep(cleanTimeout);
-    }
 }
 
 class CMainCleanup
